+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/2.0.0/).
 
 - a new progress bar in book detail screen to show local session progress
 - tap the progress slider's end duration label (full player) to toggle between total and remaining time
+- miniplayer subtitle is now configurable: show remaining time, current position, or both
 
 ## [v0.7.2] - 2026-08-30